Maintain the previous rpi provider. This is just a reference for Hydra0. Maintains better backward compatibility
