+++ lib/omniauth-rpi.rb
@@ -1,3 +1,4 @@
 require 'omniauth-rpi/version' # rubocop:disable Style/FileName
-require 'omniauth/strategies/rpi/hydra0'
-require 'omniauth/strategies/rpi/hydra1'
+require 'omniauth/strategies/hydra0'
+require 'omniauth/strategies/hydra1'
+require 'omniauth/strategies/rpi'
