+++ lib/omniauth/strategies/rpi.rb
@@ -1,0 +1,9 @@
+require 'omniauth-oauth2'
+require 'jwt'
+
+module OmniAuth::Strategies
+  class Rpi < OmniAuth::Strategies::Hydra0
+  end
+end
+
+OmniAuth.config.add_camelization 'rpi', 'Rpi'
